fix: Code scanning alert - Implicit narrowing conversion in compound assignment
Close #514

+++ src/main/java/org/cloudsimplus/hosts/network/NetworkHost.java
@@ -57,7 +57,7 @@ public class NetworkHost extends HostSimple {
     private static final Logger LOGGER = LoggerFactory.getLogger(NetworkHost.class.getSimpleName());
 
     @Getter
-    private int totalDataTransferBytes;
+    private long totalDataTransferBytes;
 
     /**
      * A buffer of packets to send for VMs inside this Host.
